+++ src/main/java/org/geowe/service/geometry/engine/JTSGeoEngineerHelper.java
@@ -21,7 +21,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import org.geowe.service.geometry.FlatGeometryBuilder;
 import org.geowe.service.model.FlatGeometry;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.spi.ReaderException;
@@ -108,28 +107,6 @@ public class JTSGeoEngineerHelper {
 		return intersects;
 	}
 
-	public FlatGeometry getIntersectedFlatGeomtetry(String wkt, Set<FlatGeometry> sourceFlatGeometries,
-			double tolerance) {
-		FlatGeometry intersectedFlatGeom = new FlatGeometryBuilder().wkt(wkt).build();
-		for (FlatGeometry flatGeometry : sourceFlatGeometries) {
-			if (getGeom(wkt).intersects(getGeom(flatGeometry.getWkt()).buffer(tolerance))) {
-				intersectedFlatGeom.setCrs(flatGeometry.getCrs());
-				intersectedFlatGeom.setId(flatGeometry.getId());
-				break;
-			}
-		}
-		return intersectedFlatGeom;
-	}
-	
-	public List<FlatGeometry> getFilledFlatGeometries(Set<FlatGeometry> sourceFlatGeometries,
-			List<String> intersectedWkts, double tolerance) {
-		List<FlatGeometry> fGeoms = new ArrayList<FlatGeometry>();
-
-		for (String intersectedWkt : intersectedWkts) {
-			fGeoms.add(getIntersectedFlatGeomtetry(intersectedWkt, sourceFlatGeometries, tolerance));
-		}
-		return fGeoms;
-	}
 
 	public Set<String> getWkts(List<FlatGeometry> overlapedUnionFlatGeometries) {
 		Set<String> wkts = new HashSet<String>();
@@ -141,8 +118,10 @@ public class JTSGeoEngineerHelper {
 
 	 public Geometry splitPolygon(Geometry poly, Geometry line) {
 	      Geometry nodedLinework = poly.getBoundary().union(line);
-	      Geometry polys = new GeometryExtractor().polygonize(nodedLinework);
-
+	      GeometryExtractor geometryExtractor = new GeometryExtractor(); 
+	      Collection<Geometry> geometries = geometryExtractor.polygonize(nodedLinework);
+	      Geometry polys = geometryExtractor.createGeometryCollection(geometries);
+	      
 	      // Only keep polygons which are inside the input
 	      List<Geometry> output = new ArrayList<Geometry>();
 	      for (int i = 0; i < polys.getNumGeometries(); i++) {