+++ src/test/java/org/geowe/service/geometry/test/AllGeometryTests.java
@@ -5,8 +5,9 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({ BufferTest.class, IntersectionTest.class, UnionTest.class, CentroidTest.class, EnvelopeTest.class,
-		DifferenceTest.class, SymDifferenceTest.class, IntersectTest.class })
+@SuiteClasses({ BufferTest.class, IntersectionTest.class, UnionTest.class,
+	CentroidTest.class, EnvelopeTest.class,	DifferenceTest.class,
+	SymDifferenceTest.class, IntersectTest.class, UnionTest.class, DivisionTest.class })
 public class AllGeometryTests {
 
 }